requirements.txt
dill == 0.3.5.1
importlib-resources == 5.9.0
matplotlib == 3.5.3
numpy == 1.23.2
omero-py == 5.11.2
pytest == 7.1.2
scikit-image == 0.19.3
scikit-learn == 1.1.2
scipy == 1.9.0

